+++ R/adjust_uc_em.R
@@ -1,5 +1,43 @@
 #' Adust for uncontrolled confounding and exposure misclassification.
 #'
+#' @description
+#' `r lifecycle::badge("deprecated")`
+#'
+#' `adjust_uc_emc()` was renamed to `adjust_uc_em()`
+#' @keywords internal
+#'
+#' @export
+adjust_uc_emc <- function(
+    data,
+    exposure,
+    outcome,
+    confounders = NULL,
+    u_model_coefs = NULL,
+    x_model_coefs = NULL,
+    x1u0_model_coefs = NULL,
+    x0u1_model_coefs = NULL,
+    x1u1_model_coefs = NULL,
+    level = 0.95) {
+  lifecycle::deprecate_warn(
+    "1.5.3", "adjust_uc_emc()", "adjust_uc_em()"
+  )
+  adjust_uc_em(
+    data,
+    exposure,
+    outcome,
+    confounders,
+    u_model_coefs,
+    x_model_coefs,
+    x1u0_model_coefs,
+    x0u1_model_coefs,
+    x1u1_model_coefs,
+    level
+  )
+}
+
+
+#' Adust for uncontrolled confounding and exposure misclassification.
+#'
 #' `adjust_uc_em` returns the exposure-outcome odds ratio and confidence
 #' interval, adjusted for uncontrolled confounding and exposure
 #' misclassificaiton. Two different options for the bias parameters are
@@ -106,10 +144,10 @@ adjust_uc_em <- function(
   xstar <- data[, exposure]
   y <- data[, outcome]
 
-  if (sum(xstar %in% c(0, 1)) != n) {
+  if (!all(xstar %in% 0:1)) {
     stop("Exposure must be a binary integer.")
   }
-  if (sum(y %in% c(0, 1)) == n) {
+  if (all(y %in% 0:1)) {
     y_binary <- TRUE
   } else {
     y_binary <- FALSE