+++ R/adjust_uc_om.R
@@ -1,5 +1,44 @@
 #' Adust for uncontrolled confounding and outcome misclassification.
 #'
+#' @description
+#' `r lifecycle::badge("deprecated")`
+#'
+#' `adjust_uc_omc_sel()` was renamed to `adjust_uc_om_sel()`
+#' @keywords internal
+#'
+#' @export
+adjust_uc_omc <- function(
+    data,
+    exposure,
+    outcome,
+    confounders = NULL,
+    u_model_coefs = NULL,
+    y_model_coefs = NULL,
+    u0y1_model_coefs = NULL,
+    u1y0_model_coefs = NULL,
+    u1y1_model_coefs = NULL,
+    level = 0.95) {
+  lifecycle::deprecate_warn(
+    "1.5.3", "adjust_uc_omc()", "adjust_uc_om()"
+  )
+  adjust_uc_om(
+    data,
+    exposure,
+    outcome,
+    confounders,
+    u_model_coefs,
+    y_model_coefs,
+    u0y1_model_coefs,
+    u1y0_model_coefs,
+    u1y1_model_coefs,
+    level
+  )
+}
+
+
+
+#' Adust for uncontrolled confounding and outcome misclassification.
+#'
 #' `adjust_uc_om` returns the exposure-outcome odds ratio and confidence
 #' interval, adjusted for uncontrolled confounding and outcome
 #' misclassificaiton. Two different options for the bias parameters are
@@ -109,7 +148,7 @@ adjust_uc_om <- function(
   x <- data[, exposure]
   ystar <- data[, outcome]
 
-  if (sum(ystar %in% c(0, 1)) != n) {
+  if (!all(ystar %in% 0:1)) {
     stop("Outcome must be a binary integer.")
   }