+++ R/adjust_uc_om_sel.R
@@ -1,6 +1,48 @@
 #' Adust for uncontrolled confounding, outcome misclassification, and selection
 #' bias.
 #'
+#' @description
+#' `r lifecycle::badge("deprecated")`
+#'
+#' `adjust_uc_omc_sel()` was renamed to `adjust_uc_om_sel()`
+#' @keywords internal
+#'
+#' @export
+adjust_uc_omc_sel <- function(
+    data,
+    exposure,
+    outcome,
+    confounders = NULL,
+    u_model_coefs = NULL,
+    y_model_coefs = NULL,
+    u0y1_model_coefs = NULL,
+    u1y0_model_coefs = NULL,
+    u1y1_model_coefs = NULL,
+    s_model_coefs,
+    level = 0.95) {
+  lifecycle::deprecate_warn(
+    "1.5.3", "adjust_uc_omc_sel()", "adjust_uc_om_sel()"
+  )
+  adjust_uc_om_sel(
+    data,
+    exposure,
+    outcome,
+    confounders,
+    u_model_coefs,
+    y_model_coefs,
+    u0y1_model_coefs,
+    u1y0_model_coefs,
+    u1y1_model_coefs,
+    s_model_coefs,
+    level
+  )
+}
+
+
+
+#' Adust for uncontrolled confounding, outcome misclassification, and selection
+#' bias.
+#'
 #' `adjust_uc_om_sel` returns the exposure-outcome odds ratio and
 #' confidence interval, adjusted for uncontrolled confounding, outcome
 #' misclassificaiton, and selection bias. Two different options for the bias
@@ -125,7 +167,7 @@ adjust_uc_om_sel <- function(
   x <- data[, exposure]
   ystar <- data[, outcome]
 
-  if (sum(ystar %in% c(0, 1)) != n) {
+  if (!all(ystar %in% 0:1)) {
     stop("Outcome must be a binary integer.")
   }